changed code to remove service layer
